Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Do you need the 26/27 changes?

Do you mean the ignore.cil file changes? My understanding was that you had to tell the compiler that the API wasn't present, but I'm not overly familiar with the builds. Is there a specific reason we wouldn't need it?

Patch-set: 4
Reviewer: Gerrit User 1385031 <1385031@85c56323-6fa9-3386-8a01-6480fb634889>