Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Why is this needed for perfetto when nothing else requires it?  Is perfetto doing something different?

We execute in a non-shell domain when run from shell to get the same results as when we get ran in other contexts.

We daemonize but want to allow the user to send SIGINT end the trace early but gracefully.

Is there any other way we could to this?

Patch-set: 2
Reviewer: Gerrit User 1268571 <1268571@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0