Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> What exactly is this CL supposed to do?  It's not granting any new permissions.

These patch needs to add the hal_bootctrl to the white-list of sys_rawio.
Because we need to add the sys_rawio permission for hal_bootctrl to do boot ctrl.

Without this patch, we can not build pass with the modification of sepolicy.

Patch-set: 6
Reviewer: Gerrit User 1248931 <1248931@85c56323-6fa9-3386-8a01-6480fb634889>