Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Can you revert aosp/764971 and ag/5189359, which added bufferhub binder service, instead of creating new changes?

I don't know if it's generally preferred to revert a CL for non-bug related reason. Let's see if Steven and Jeffrey have any ideas.

I just tried to revert the CL and get a merge conflict (see 846457, I don't know how the short link works here it's obviously not aosp/CL number) since this is not a pure revert. A line of comment added in the original CL should still be there (or just be modified), and some lines in compat/ are already removed in CL 774568.

Patch-set: 1
Reviewer: Gerrit User 1370074 <1370074@85c56323-6fa9-3386-8a01-6480fb634889>