Update patch set 2
Patch Set 2:

AIUI and according to:
https://source.android.com/security/selinux/images/SELinux_Treble.pdf

private/ is for Android internal things that aren't guaranteed to be stable. It makes sense for private/ then since the loader's policies are Android internal. But I will defer to the security team's expert opinion on this matter.

Patch-set: 2
Reviewer: Gerrit User 1124848 <1124848@85c56323-6fa9-3386-8a01-6480fb634889>