Update patch set 1
Abandoned

Not sure why this was bumped (or why I hadn't abandoned earlier), but there aren't any current or planned users of this, so abandoning.

If you have a patch that would make use of it, could you post that too?  Otherwise I'll abandon this to prevent further confusion.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon