Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Is CTS the only user of this property?

Yes, currently that's our only plan for it.   We need a way for CTS to know that the ANGLE package is present for devices that upgrade to Q (it's required for new devices shipping with Q).

This was our best idea for a way for CTS to query if ANGLE is present without needing to know the name of the ANGLE package, since it can be anything the OEM wants (com.google.android.angle, com.android.angle, com.huawei.android.angle, etc.).

Is this an appropriate use of one of these properties?

Patch-set: 1
Reviewer: Gerrit User 1368430 <1368430@85c56323-6fa9-3386-8a01-6480fb634889>