Update patch set 1
Patch Set 1:

Suggest we remove all/most of the liblog code that supports runtime registration of event log tags and/or maybe is_loggable filtration upstream of the logd socket.

[TL;DR]

... and accept that we will never solve the build problem that it was created to address. ccross@'s goal was to remove all static event log tags, and without an owner to complete this project and deprecate static event log tags, most of the code is not currently being used. It should be noted the runtime registration of event log tags is unused _only_ because the feature request was not quite complete by the time I stopped maintaining the logger and a new owner was not found for the feature.

Recommend keeping the code in logd since some of it is intertwined with the per log-tag statistics used to point to spam violators; just remove the code in liblog.

Unless we keep this feature and move forward there are design issues that prevent your request:

logcat in every context needs to at least try to access the contents.

liblog in every context needs to at least try to access the file for is_loggable determination.

The problem was it would take 200us round-trip to ask logd for entries in the file. If the file was directly visible, it would take on slower devices from 2 years ago 30ns (500ns first time to map it in) to get the value, and log messages have to be restricted to less than 20us _total_ overhead. The reason access was restricted to the file, was the content could provide clues as to which execution paths have been run (jeffv@'s objection if memory serves).

We could remove liblog checking is_loggable to remove the accesses, and move that overhead into logd. The disadvantage is that is_loggable has been used to pre-filter high noise sources (blacklist) and if the filtration is only in logd, the socket can be DOS'd without any recourse inside liblog, which with direct access to the file is very efficient.

This would also make logcat much slower as it may need to check with logd on every event it printed instead of accessing the file; but it is not as time-sensitive and can maintain a cache if necessary.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>