init can call setns
Init creates two mount namespaces one for pre-apexd processes and the
other for post-apexd processes. That is to have different
mounts for the bionic libs and linker for the two types of processes.
For example., for pre-apexd processes
/system/lib/libc.so is a mount point for /system/lib/bootstrap/libc.so

for post-apexd processes
/system/lib/libc.so is a mount point for
/apex/com.android.runtime/lib/bionic/libc.so

Using unshare and setns, init first starts with the mount namespace for
the pre-apexd and then switches to the other mount namespace when APEXes
are ready. It then occasionally switches to pre-apexd mount namespace
when it has to re-launch a pre-apexd process (e.g. the process has
crashed, etc.)

Bug: 120266448
Test: m device boots

Change-Id: I5da2d4c4ad6a5579d38ef17094e4958b39307a5a