Update patch set 1
Patch Set 1: Code-Review+1

this adjusts a neverallow rule, which is itself a CTS test. This CTS test relaxation is needed to allow newer kernels to run. This change will not induce failures on previously passing devices - this is a strict relaxation of a CTS test.

+jgalenson for owners +1, as I am not in the OWNERS file for pie-cts-dev.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1