Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> This LGTM for now to make the existing tests work, but I'd like to revisit whether this link is truly necessary.

Because to remove the symbol link with the alternative solutions that also needs to store storage something in storage and needs design review or go/wipeout, it needs to solve the storage problems.

The DocumentsUI using link is not truly necessary after code review. The symbolic link absolutely needs to remove and fire b/123666046 to trace.

Patch-set: 1