Update patch set 1
Patch Set 1: Code-Review-2

> Patch Set 1: Code-Review+1
> 
> Since the authsecret HAL is available on Pie, this should be backported to the next 9.0 r_X branch as well.
> 
> Btw, there are lots of other bumped HAL versions not covered, e.g. the power HAL available in Pie is on v1.3, whereas the sepolicy only has 1.0 covered. Might be worth looking into.

Thank you for your change, but I don't think we should merge it. This specific implementation is a mock (doesn't do anything and shouldn't be shipped in production). vendor/file_contexts here is mostly just for default/shims, and most devices should have custom implementations and specify the sepolicy of these implementations with BOARD_SEPOLICY_DIRS. When a device has a custom implementation, it should be in files related to that device and for source compatibility, it should be named something specific which does not conflict with AOSP. For instance, you could name something android.hardware.authsecret@1.0-service.foo for a device named foo. In this way, forking code is avoided (and the HIDL interface itself provides a stable ABI).

Patch-set: 1
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2