Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (1 comment)
> 
> I think the simplest and least fragile way to implement this is to create a separate shell script. e.g. /system/bin/wake_traceur.sh. Then have perfetto or traced call that directly with a domain transition.

We tried that here (in this intermediate patchset):
https://android-review.googlesource.com/c/platform/external/perfetto/+/889213/3
and couldn't get it to work, but maybe we were missing something with the domain transition. Would we need to do something different in notify_traceur.cc?

Note: I'm becoming anxious about getting this squared away by the IC extension deadline tomorrow so if it could be an option to reassess/rework the separate script strategy in a follow-up change, that would be appreciated! Understood if not though.

Patch-set: 5
Reviewer: Gerrit User 1170070 <1170070@85c56323-6fa9-3386-8a01-6480fb634889>