Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > jeffv, the problem with a dedicated script is coupling. The script can't just send an implicit broadcast, since that won't wake Traceur post-O. Toggling a property lets Perfetto just indicate generally that a trace has ended, and Traceur can (without Perfetto's help) register a bit of init code to watch for this property to change and run some custom application logic (e.g., sending an actual intent) instead.
> 
> my suggestion would do the same. traceur would own the executable, and perfetto (or traced) would just know the path. So path string instead of property string. It really doesn't make a huge difference to me.

It's a big difference. The property acts as a rendezvous point, whereas a script filename doesn't. How does traced know _which_ script to run when a particular trace ends? Either the script is hardcoded or it's client-supplied. If the script is hardcoded, then we create an ugly tight coupling between traced and its clients. If the script is client-supplied, that's remote code execution, since the client is basically telling traced to arbitrary work with *traced* credentials.

If we set a property instead of running a script, the wiring is dynamic, because we accumulate rc files during build, rather than being hardcoded into traceur and traced. See what I mean?

Patch-set: 4