Update patch set 1
Patch Set 1:

This is probably fine to merge here (so that the prebuilts are accurate), but it won't hurt anything. We should be careful though and make sure that this patch is in whatever release branch though (for instance, it might need to be cherry-picked to the GSI branch).

Patch-set: 1
Reviewer: Gerrit User 1124693 <1124693@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1061057 <1061057@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>