Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+2
> 
> Please make sure you've checked with privacy about the data we persist to these directories. Persisting stats on the device leaks past device activity, which may or may not be a privacy concern. Historically, we've told people that phone log data is temporary and erased on reboot (https://android-developers.googleblog.com/2010/08/best-practices-for-handling-android.html , last paragraph)
> 
> Please make sure you've checked with the device health team about the data persisted to these directories. Excessive flash writes can prematurely wear down flash, causing the device to become slower over time. Inadvertent or intentional excess logging can consume storage space, removing space from other applications.

Thanks for the pointers, Nick. The cl was on auto submit so it got merged before I had the chance to do all that you suggested.
By the nature of this cl, it writes once on boot and once on shutdown, which is less than what we are already doing.
The data is internal logic so no privacy concern should come from it.

Patch-set: 3