sepolicy change for NetworkStack package changes
The NetworkStack may have different package names depending on the
system, and may not be signed with the platform key.
Do not hardcode the package name, and instead update the sepolicy
rule to apply context based on UID (which is fixed regardless of the
NetworkStack package used), and priv-app status.

Test: built, flashed, verified NetworkStack working properly with
      different package name
Bug: 124033493

Change-Id: Iec83b12df2762a3339e1e172f3790a5d931169e4