Update patch set 2
Patch Set 2: Presubmit-Ready+1

Hi Jeff,

Here is a change to stop hard-coding the package name for the NetworkStack, since on devices shipping the mainline module, it will not be signed with the platform cert and may not have the same package name. The context is now set based on the UID (which is pinned) and isPrivApp.
Let me know if you have any concerns with the change.

Patch-set: 2
Reviewer: Gerrit User 1054468 <1054468@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1288034 <1288034@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1