Game Driver: sepolicy update for plumbing GpuStats into GpuService
All the non-system and non-privileged app process will send GPU status
to the GpuService during the GraphicsEnvironment setup stage for the
process. untrusted apps process need to be able to communicate with
GpuService as well, otherwise it will be blocked for a few seconds for
failing to get GpuService at app start up.

Bug: 123529932
Test: Build, flash and boot. No selinux denials.
Change-Id: Ic7687dac3c8a3ea43fa744a6ae8a45716951c4df