Update patch set 6
Patch Set 6:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > > Patch Set 5:
> > > 
> > > > Patch Set 5:
> > > > 
> > > > So, in system_server, you're checking to see if you're running a GSI image by seeing if a file exists? Can we do this with a system property instead? Have init set a property if it's loading GSI instead of the normal system image?
> > > 
> > > Is there a reason why you're recommending the property approach vs the current approach. Setting a property would involve a round trip through the property server, which increases dependencies, increases the amount of state that gsi needs to maintain, and as a result, increases the probability of failures.
> > 
> > If that approach is broken then all of Android is broken. I don't understand the rest of your argument. "increases the amount of state that gsi needs to maintain"? It's a single property, that can potentially be used by multiple components to determine if they're running in a GSI setup. i.e. it scales much better than granting components access to data that they otherwise do not need access to. 
> > 
> > Having system_server check for files in a location seems much more fragile and potentially racy (do these files get deleted, modified, etc?) than a property that init sets before even loading the GSI image.
> 
> It does not get deleted or modified (libfstab relies on this).
> 
> But honestly, a property sounds fine to me, as long as (1) it's set very early in the boot sequence so there's no race with zygote, and (2) can only be accessed by system_server. If that's preferable, I can revive the original CL which added a gsid_prop label.

(chiming in late)

I still don't get why waiting for gsid is a problem. We are creating a side-channel API (with property or with the file check). The "implementation details" of _how_ GSIs are managed are all within 'gsid'. So it makes sense to ask GSID if we are running the GSI.

Ofcourse, metadata files are now forever stable for gsid cause we can't change them (otherwise, one gsid won't be able to talk to the other instance of gsid :D), which makes all this slightly better. 

It still _feels_ wrong to bypass gsid for this. Properties are better than direct file access in that regard, cause the property is also a known API we've used in the past a lot.

Patch-set: 6
Reviewer: Gerrit User 1108565 <1108565@85c56323-6fa9-3386-8a01-6480fb634889>