Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Is this still needed?  It looks to me like the (temporary) revert from this morning means this isn't actively happening at the moment.

Actually, commit mentioned in http://b/124345887#comment10 wouldn't have caused this. This is a known issue and I added a entry in bug_map to track it earlier (http://aosp/799893) but the issue now is for some reason, the denial has a different scontext, so that's why adding a new entry.

Patch-set: 2
Reviewer: Gerrit User 1101746 <1101746@85c56323-6fa9-3386-8a01-6480fb634889>