move untrusted app access for /proc/net to private policy
This is a no-op for pi-dev. The same permissions are granted, they're
just moved from the vendor partition to the system partition.

This ensures that when these permissions are removed in a future version
of Android they are not re-granted in vendor policy on Treble/GSI
builds.

Bug: 123616769
Test: testTcpDefaultRwndSane testProcNetSane
Change-Id: I9cba32608cdeb2158d20f522c32cf6284ab34c95
Merged-In: I2304e3e98c0d637af78a361569466aa2fbe79fa0