Allow apexd to link apex_data_file.
This is needed in cases SELinux labels are restored under /data/apex by
an external process calling restorecon. In normal condition files under
/data/apex/active retain the label staging_data_file used at their
original creation by StagingManager. However, we observed that the label
might be changed to apex_data_file, which we were able to reproduce by
running restorecon.

In that case, in order to prepare the /data/apex/backup directory before
applying a new set of APEX files, we need the link permission, which is
being added in this CL.

We theorized that the unlink permission would have been needed to for
when files under /data/apex/backup are removed after a successful
installation, however we weren't able to trigger any violation in our
tests, and therefore we are not adding the unlink permission yet.

This CL also remove some stale rules being addressed since.

Test: ran restorecon on files in /data/apex/active, attempted installing
a new apex which triggered the violation when files are linked to
/data/apex/backup. With this CL, the operation succeeds.
Bug: 112669193
Change-Id: Ib4136e9b9f4993a5b7e02aade8f5c5e300a7793c