Update patch set 2
Patch Set 2:

The change in system/apex is sort of dependent on this one but this change is independent of the system/apex change. Basically all keys on /product or /system/products are unlabelled and this upload should take care of that.

Patch-set: 2
Reviewer: Gerrit User 1226510 <1226510@85c56323-6fa9-3386-8a01-6480fb634889>