Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Is there a companion CL which uses these permissions?

Not yet. Shall I do it when it really becomes necessary? Marco, do you want chime in?

> 
> Some initial impressions:
> 
> 1) I thought the list of APEXes for a particular release were fixed? If so, the dependent components should already know by name the APEXes they intend to use. From that regard, this change seems unnecessary.

Yes, it's fixed. But I think hardcoding the names might not always be a good practice.

> 
> 2) Probing of installed components is already possible without this change. A component exists if "access('/path/to/apex', F_OK)" returns true. This does not require directory read permissions. The difference between directory reads and probing is that directory reads allow you to enumerate ALL installed APEXes, where probing allows you to determine the existence of a particular APEX without knowing all the other APEXes.

Yes, you are right. So this change is actually to being able to enumerate all APEXes in order not to hardcode their names.

> 
> 3) Enumeration instead of probing often leads to false architectural dependencies. It appears, from device activity and reading security policy, that one component would depend on every other component. These false dependencies make it harder to understand what's happening on the system and inhibit refactoring. Probing avoids those problems and makes the dependency graph clearer.

Maybe I am missing something from your argument. But I have thought enumerating loosens the dependency as a client of an APEX doesn't need to rely on the APEX name. Instead, the client can search for an APEX that satisfies certain condition - much like app activities are searched via intents, not by their names.

> 
> 4) Allowing this to "domain" (all processes) is overly broad. Not only does this lead to false architectural dependencies (see #3), the ability to list these directories becomes a defacto API available to any Android app or Treble component, with corresponding ongoing support requirements.

Hmm, I have thought this would be okay since 'domain' already has access to other similar directories like /system/app. And I don't quite follow your argument about the defacto API. It *is* true that they becomes the defacto API, but I think that's not a problem (or that's at least a KI). For example, consider a library path /system/lib/libFoo.so where libFoo.so is a private system library. Although it is a private system library, apps can open/read it and thus will break if we remove the library or change the path in the next release. But we don't call it a compatibility issue (do we?).

> 
> Thoughts?

Patch-set: 1
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0