Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Is there a companion CL which uses these permissions?
> 
> Not yet. Shall I do it when it really becomes necessary? Marco, do you want chime in?

Having a concrete CL would help frame the discussion and avoid us getting stuck discussing abstracts. I'd prefer to wait until a CL is available if feasible.

> 
> > 
> > Some initial impressions:
> > 
> > 1) I thought the list of APEXes for a particular release were fixed? If so, the dependent components should already know by name the APEXes they intend to use. From that regard, this change seems unnecessary.
> 
> Yes, it's fixed. But I think hardcoding the names might not always be a good practice.

Because the list of APEXes are fixed per release, the mapping from criteria to APEX is also fixed. If the goal was to avoid hardcoding these names, supporting a configuration file which mapped criteria to APEX name might be a cleaner, faster way to have the same effect. The client would then read the configuration file and use the APEXes specified in that file, without exposing or accessing unnecessarily other, unrelated APEXes.

> 
> > 
> > 2) Probing of installed components is already possible without this change. A component exists if "access('/path/to/apex', F_OK)" returns true. This does not require directory read permissions. The difference between directory reads and probing is that directory reads allow you to enumerate ALL installed APEXes, where probing allows you to determine the existence of a particular APEX without knowing all the other APEXes.
> 
> Yes, you are right. So this change is actually to being able to enumerate all APEXes in order not to hardcode their names.
> 
> > 
> > 3) Enumeration instead of probing often leads to false architectural dependencies. It appears, from device activity and reading security policy, that one component would depend on every other component. These false dependencies make it harder to understand what's happening on the system and inhibit refactoring. Probing avoids those problems and makes the dependency graph clearer.
> 
> Maybe I am missing something from your argument. But I have thought enumerating loosens the dependency as a client of an APEX doesn't need to rely on the APEX name. Instead, the client can search for an APEX that satisfies certain condition - much like app activities are searched via intents, not by their names.

The loosening of dependencies is what I'm trying to avoid. Enumeration would require the client to query the functionality of *all* the APEXes on the system, even though the functionality is knowable at build time. Having an explicit dependency graph architecturally helps understand the communication flows of the system, and the security policy is intended to help build this communication dependency graph.

> 
> > 
> > 4) Allowing this to "domain" (all processes) is overly broad. Not only does this lead to false architectural dependencies (see #3), the ability to list these directories becomes a defacto API available to any Android app or Treble component, with corresponding ongoing support requirements.
> 
> Hmm, I have thought this would be okay since 'domain' already has access to other similar directories like /system/app. 

Access to /system/app and other /system files is a historic artifact, and was primarily allowed because it aided SELinux bringup. Additionally, at the time it was allowed, there was more of a focus on enforcing pure security goals, not Android architectural goals. 

> And I don't quite follow your argument about the defacto API. It *is* true that they becomes the defacto API, but I think that's not a problem (or that's at least a KI). For example, consider a library path /system/lib/libFoo.so where libFoo.so is a private system library. Although it is a private system library, apps can open/read it and thus will break if we remove the library or change the path in the next release. But we don't call it a compatibility issue (do we?).

Actually, yes, this is a compatibility issue which has cost us huge amounts of engineering time. This was one of the primary reasons Android introduced linker namespaces and we've started being more strict about what is exposed on the /system partition. If we can't enforce visibility control on things like /system/lib/libFoo.so, then changes cannot be made to that library without a comprehensive review of every app. We've done similar things for apps accessing private java APIs, eg go/hiddenapi-technical-doc . This has consumed A LOT of engineering time.

The one I've been personally fighting has been the Chase banking app listing out all the files in /system/bin, see b/119605322 and https://android-review.googlesource.com/c/platform/system/core/+/822955 for background. The over exposure of files in /system/bin has inhibited security tightening and introduces false architectural dependencies. 



> 
> > 
> > Thoughts?

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>