Update patch set 1
Patch Set 1:

Nick, thank you very much for sharing the opinions. I agree with you that the same could be achieved by having a config file having the APEX names. So, I am abandoning this.

-----
One nit: the linker namespace cannot completely hide a library. It's existence isn't hidden. Apps are just not allowed to load it using the system dynamic linker. Since apps can still open/read the file, they could load it using their own linker (if any), which would be much annoying to them but anyway possible (and I saw some).

So my point is that, there are certain pieces of information that are inevitably exposed to apps (e.g. offset between two symbols, size of a file, etc.) which we don't want to call as API but changing might break a few exceptional apps that have been relying on the pieces (intentionally or not). The corollary of this is that the best strategy for app compatibility is doing nothing, which of course does not make sense. I don't know where does the line should be. :)

Patch-set: 1
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0