Update patch set 1
Patch Set 1:

the concrete use case is the media extractor process wanting to load all the available extractors on the device when it starts up.

Currently we look in /system/lib[64]/extractors and /apex/com.android.media/lib[64]/extractors, but if we could read the directories under /apex, we could look for extractors in every mounted apex. This would let vendors package their extractors as an apex. It would also give us more flexibility in the way we distribute things.

Patch-set: 1