Update patch set 2
Patch Set 2:

Nick, let me know if you have any concerns with this; happy to setup a meeting to talk more.

The bottom line is that vold implements filesystem checkpointing, and apexd changes its behavior based on whether filesystem checkpointing is supported. It also needs to be able to tell the checkpointing code to abort the current checkpoint (eg, if an APEX is bad).

Patch-set: 2
Reviewer: Gerrit User 1013030 <1013030@85c56323-6fa9-3386-8a01-6480fb634889>