Update patch set 4
Patch Set 4:

(2 comments)

> Patch Set 2:
> 
> (2 comments)
> 
> I think this is fine. It's not immediately clear to me that there's a huge security advantage from keeping apexd away from vold.
> 
> +jsharkey for a second opinion. In the old days, only system_server could talk to vold. Access to vold has been expanding a lot (for good reasons), but it's getting to the point where I'm starting to get uncomfortable. Nothing actionable for this CL, just me talking out loud.

One thing we could decide to do is to put the checkpoint APIs in a separate binder service. That service would then still be implemented by vold, but we can then enforce that apexd can only retrieve the checkpoint service, not the regular vold service.

Of course, binder calls are still only limited by domain, so if apexd were to get a reference to the vold binder service by other means outside of servicemanager, we can't stop it from talking from it. But this does seem like an improvement, and perhaps something we can track for R? If this makes sense to you I'll file a bug.

Patch-set: 4
Reviewer: Gerrit User 1013030 <1013030@85c56323-6fa9-3386-8a01-6480fb634889>